Resolve storage inside ProgressiveSinkOpExec (#3146)
In the previous implementation, the storage object was resolved and
assigned directly within the descriptor. This caused issues during
compilation, as the storage object in this setup was not serializable.

This PR introduces a change in the lifecycle management of storage
resolution. The descriptor (logical layer) now only records the upstream
operator’s ID. The actual storage object is resolved later, within the
ProgressiveSinkOpExec, using the upstream operator’s ID.

To facilitate resolving storage objects inside ProgressiveSinkOpExec, a
`ResultStorage` singleton has been introduced. This singleton maintains
a mapping of all workflows’ `OpResultStorage` objects against their
workflow IDs. However, this approach has one limitation: the singleton
is only available on the master node. As a result, all sink executors
must run on the master node. While this is consistent with the current
design, in the future, we should address this by moving the storage
mappings to an external, distributed, and persistent location.