Address Result Panel Getting Sticky Too Easily (#3136)
## Purpose
Address #3042
There was an issue where visualization results were very easy to make
panels "sticky", meaning the panels would be dragged even though the
user is no longer holding down the mouse. This PR aims to fix this
issue.
## Changes
Add logic to modify the z-index of visualization results directly. If
the panel is being dragged, the visualization result's z-index becomes
-1, and if the panel is not dragged, the z-index changes back to 0 so
that the user can interact with the visualization.
## Demo
Before

![chrome_svDxLZYYvl](https://github.com/user-attachments/assets/8299e19d-9c66-4d9c-a40f-3fed8f7823be)
After

![R3oNSHxH8J](https://github.com/user-attachments/assets/65069f65-3995-435a-88f6-d7cc5b66ddf1)

---------

Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>