Add Cost Estimator Using Past Statistics for Schedule Generator (#3156)
#### This PR introduces the `CostEstimator` trait which estimates the
cost of a region, given some resource units.
- The cost estimator is used by `CostBasedScheduleGenerator` to
calculate the cost of a schedule during search.
- Currently we only consider one type of schedule for each region plan,
which is a total order of the regions. The cost of the schedule (and
also the cost of the region plan) is thus the summation of the cost of
each region.
- The resource units are currently passed as placeholders because we
assume a region will have all the resources when doing the estimation.
The units may be used in the future if we consider different methods of
schedule-generation. For example, if we allow two regions to run
concurrently, the units will be split in half for each region.

#### A `DefaultCostEstimator` implementation is also added, which uses
past execution statistics to estimate the wall-clock runtime of a
region:
- The runtime of each region is represented by the runtime of its
longest-running operator.
- The runtime of operators are estimated using the statistics from the
**latest successful execution** of the workflow.
- If such statistics do not exist (e.g., if it is the first execution,
or if past executions all failed), we fall back to using number of
materialized edges as the cost.
- Added test cases using mock mysql data.