Disable the “Clone” button for restricted and inactive users (#3208)
### Purpose
It looks to restricted/inactive users as they can clone a workflow, when
they actually cannot. To reduce the confusion, the clone button is
disabled in the GUI.
fix #3066 

### Changes

Added variables isAdminOrRegularUser in
hub-workflow-detail-component.ts. Check the user role. If the user is
not an admin or regular user, disable the clone button.



https://github.com/user-attachments/assets/c426c50d-fb31-4377-a2b7-3de2a24da512

---------

Co-authored-by: Texera <texera@MacBook-Air.local>