Add Network Graph operator for visualizing graph structured data (#3207)
### Purpose
Add Network Graph operator to visualize graph structured data and
network topologies. The operator gets a table as an input, where one
column is specified as source and another as destination. Each entry in
a column is a node and the source and destination entry in a row form an
edge.

### Changes
Added NetworkGraphOpDesc
Added Icon for operator

CSV file used for testing: 

[musae_git_edges.csv](https://github.com/user-attachments/files/18388032/musae_git_edges.csv)



https://github.com/user-attachments/assets/9a41a7ab-27fa-4f38-98ec-958874ca08fe

---------

---------

Co-authored-by: Texera <texera@MacBook-Air.local>
Co-authored-by: Texera <texera@dhcp-v062-167.mobile.uci.edu>
Co-authored-by: GspikeHalo <109092664+GspikeHalo@users.noreply.github.com>
Co-authored-by: gspikehalo <2318002579@qq.com>