Fix download operators result (#3241)
This PR is in continuous of [PR
3221](https://github.com/Texera/texera/pull/3221) to implement download
logic in backend.

## Export request structure updates
In this PR, we update `ResultExportRequest` to include a new variable
named `destination`. If its set to `local`, means the user wants to
download operators result. Also, `operatorIds` is now plural so with one
request, multiple operators can be exported (locally or to dataset).

## Export endpoint updates
In this PR, we update `ResultResource` to support `destination=local`.
In this regard, we have two cases:
- If destination is local and one operator is selected, download one
operator result based on export type
- If destination is local and multiple operators are selected, download
a zip file containing all
The export to dataset follows the old logic.

## Export service updates
In this PR, we update `ResultExportService` to have two core more new
functions:
- `exportOperatorResultAsStream`: For local download of a single
operator. Streams the data directly.
- `exportOperatorsAsZip`: For local download of multiple operators. Zip
the output and stream back.

## Demo video


https://github.com/user-attachments/assets/b46b33a7-9503-4ed8-99f2-9dc15878e54d

---------

Co-authored-by: Jiadong Bai <43344272+bobbai00@users.noreply.github.com>