Improve the compilation of visualization operators by suppressing the exception to codes  (#3265)
This PR tries to improve the workflow editing experience with
visualization operators.

### Before the changes

When users drag a visualization operator out, the compilation is likely
to fail because users have not configured the properties yet and in the
compiler `generatePythonCode` is called and lots of assertion fail the
calling. This often breaks the schema propagation which is quite
unfriendly to the user.
![2025-02-13 18 34
08](https://github.com/user-attachments/assets/0273059c-a5c5-42ca-9d8a-f1c83c77a190)



### How the changes are done

This PR adds a try-catch around `generatePythonCode` function. If there
is any exception happen during the python code generation, it will be
caught. The exception will be reflected in the generated python code.

Here is the demo after the changes:
![2025-02-13 18 35
04](https://github.com/user-attachments/assets/545f4906-4f64-4d9b-be43-8c1df75a2bf6)