Bring back the dataset size calculation and display (#3320)
This PR adds the functionality on the backend to calculate the size of a
dataset, by:
1. call LakeFS api to get the latest version of a dataset
2. call LakeFS api to get the list of metadata of each object in this
version
3. Do a sum on the size

This PR also brings the frontend dataset size display back.

<img width="1466" alt="Screenshot 2025-03-12 at 5 57 09 PM"
src="https://github.com/user-attachments/assets/4543d7b6-b878-4e81-8f2d-6fde0d8bf521"
/>


Note that the cost of calculating the size of a dataset is 2 additional
http requests. In the future we should consider optimize it.