Fix Iceberg Catalog on Window Machines (#3322)
Fix the bug where the workflow throws a No storage is found for the
given URI error when a Python UDF is present, and the workflow is
running on a Windows machine.

The root cause is that the Java side writes the path as
C:/Users/texera/core to the Iceberg Catalog in the database, but PyArrow
on the Python side cannot parse it. The acceptable format for PyArrow is
C/Users/texera/core.