Refactor Operator Port Result Storage URI (#3332)
As mentioned in #3312, this PR removes `layerName` introduced in #3279
from the URIs of operator port results.

### Main Idea

We do not want to use a combination of LogicalOperatorId + LayerName +
OutputPortId (local to each physical operator) in the storage URIs and
related APIs. Instead, we want to directly use a `GlobalPortIdentity`
object.

### Serde

As GlobalPortIdentity is a protobuf object, we cannot rely on protobuf's
built-in serde methods to encode a GlobalPortIdentity in the URIs
because it is not intended to be used for URIs or file names.

The serde of GlobalPortIdentity has very strict requirements:

1. It needs to be compatible as part of a URI. (protobuf's built-in
serde method does not satisfy this)
2. It needs to be compatible with file names in different file systems.
(JSON is out as ":" does not work with Windows file names)
3. It should not contain "`_`" because we reserve it for our own logic
of parsing URIs. For example, we rely heavily on "`_`" to parse other
parts of the URI, e.g., "`eid`"
4. It needs to be human-readable. (Base64 does not satisfy this)

To satisfy these requirements, I implemented a custom serde method using
the following pattern:


`(logicalOpId=<logicalOpId>,layerName=<layerName>,portId=<portId.id>,isInternal=<portId.internal>,isInput=<input>)`

 This implementation is used for both Scala and python codebase.

### Other Refactoring

Several other refactoring / code optimizations related to port result
URIs are included in this PR. I also discovered a potential problem with
our Python implementation of sanitizeURIs, which is in consistent with
our Scala implementation. I fixed this problem in this PR.

### TODOs

- Currently all the frontend requests only include the logicalOpId and
does not have `GlobalPortId` information. This means we have to retrieve
all the URIs from either the database or `ExecutionResourceMapping` for
the requested `eid` and find one URI that matches the logicalOpId,
portId, and is an external port. This is an O(n) lookup. Once we also
include physical plan information in the frontend, we can optimize this
to O(1).
- Once user system is permanently enabled, we can remove some logic for
using `ExecutionResourceMapping` to do lookups.

## Applying DDL change

This PR modifies the `operator_port_executions` table. Similar to #3279:

Run `core/scripts/sql/updates/03.sql` in psql to apply the changes to
`operator_port_executions` table.

The data from existing `operator_port_executions` table will be removed.
It will clear your recent execution results.